LayoutTests/platform/mac/editing/mac/input/firstrectforcharacterrange-styled-expected.txt

EDITING DELEGATE: shouldBeginEditingInDOMRange:range from 0 of BODY > HTML > #document to 4 of BODY > HTML > #document
EDITING DELEGATE: webViewDidBeginEditing:WebViewDidBeginEditingNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: shouldChangeSelectedDOMRange:range from 1 of #text > I > B > BODY > HTML > #document to 1 of #text > I > B > BODY > HTML > #document toDOMRange:range from 1 of #text > I > B > BODY > HTML > #document to 3 of #text > I > B > BODY > HTML > #document affinity:NSSelectionAffinityDownstream stillSelecting:FALSE
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
EDITING DELEGATE: webViewDidChange:WebViewDidChangeNotification
EDITING DELEGATE: webViewDidChangeSelection:WebViewDidChangeSelectionNotification
123456
10,572,0,18
18,572,0,18
61,572,0,18
61,572,0,18
10,572,8,18
52,572,9,18
52,572,9,18
52,572,9,18
10,572,51,18
10,572,51,18
10,572,51,18
18,572,43,18
0,0,0,0
0,0,0,0
0,0,0,0
0,0,0,0
